+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes are documented in this file using the [Keep a CHANGELOG](htt
 
 ## 1.11.2 - Unreleased
 
+* Fixed [#488](https://github.com/NeoVintageous/NeoVintageous/issues/488): `gx` doesn't work on some markdown links
 * Fixed [#487](https://github.com/NeoVintageous/NeoVintageous/issues/487): `{` in Visual mode is off-by-one
 * Fixed [#486](https://github.com/NeoVintageous/NeoVintageous/issues/486): `}` in Visual mode is off-by-one
 * Fixed [#485](https://github.com/NeoVintageous/NeoVintageous/issues/485): `'{a-z}` should jump to the first non-blank character in the line